requirements.txt
google-generativeai>=0.7.2
google-api-python-client>=2.141.0
google-auth>=2.33.0
google-auth-oauthlib>=1.2.1
chromadb>=0.5.5
pandas>=2.2.2
openpyxl>=3.1.5
numpy>=1.26.4
tqdm>=4.66.4
fastapi
uvicorn
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
chromadb
